% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robinhoodR6.R
\docType{data}
\name{Account}
\alias{Account}
\title{Account R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Account
}
\arguments{
\item{username_}{Robinhood login username. NOT email.}

\item{password_}{Robinhood login password. Can only contain letters and numbers in plaintext.}
}
\description{
Create an instance of the Account class.
}
\keyword{datasets}
