% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robinhoodR6.R
\name{get_balances}
\alias{get_balances}
\title{Get current balances.
Retrieve current portfolio market value and positions.}
\usage{
get_balances(account)
}
\arguments{
\item{account}{Initialized instance of the R6 Account class. Generated by Account$new(login).}
}
\description{
Get current balances.
Retrieve current portfolio market value and positions.
}
