% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robinhoodR6.R
\docType{data}
\name{Orders}
\alias{Orders}
\title{Orders R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Orders
}
\arguments{
\item{account}{Initialzied instance of Account class}
}
\description{
Retrieve all transation history and profit/loss summary.
}
\keyword{datasets}
