% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{parsePosition}
\alias{parsePosition}
\title{Parse position helper method.}
\usage{
parsePosition(position)
}
\arguments{
\item{position}{position list object returned from rh_getPositions.}
}
\description{
Parses position in list structure to be used in creating a positions data.frame.
}
