% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_createAuthHeader}
\alias{rh_createAuthHeader}
\title{Create HTTP header.}
\usage{
rh_createAuthHeader(userAuthToken)
}
\arguments{
\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}
}
\description{
Creates HTTP header to pass into GET HTTP requests. Required for all authentication requests.
}
