% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_createPositionsTable}
\alias{rh_createPositionsTable}
\title{Get positions table.}
\usage{
rh_createPositionsTable(userAuthToken, accountNumber)
}
\arguments{
\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}

\item{accountNumber}{account number generated by rh_getAccountNumber.}
}
\description{
Parses current positions into a data.frame.
}
