% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_getAccountID}
\alias{rh_getAccountID}
\title{Get account ID.}
\usage{
rh_getAccountID(userAuthToken)
}
\arguments{
\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}
}
\description{
Gets unique robinhood account ID.
}
