% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_getAccountNumber}
\alias{rh_getAccountNumber}
\title{Get account number.}
\usage{
rh_getAccountNumber(userAuthToken)
}
\arguments{
\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}
}
\description{
Gets unique robinhood account number. Used in all account based requests.
}
