% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_getAuthToken}
\alias{rh_getAuthToken}
\title{get r AuthToken.}
\usage{
rh_getAuthToken(username_, password_)
}
\arguments{
\item{username_}{Robinhood login username. NOT email.}

\item{password_}{Robinhood login password. Can only contain letters and numbers in plaintext.}
}
\description{
Get oauth2 login token. Passed to all API requests.
}
