% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_getDayPL}
\alias{rh_getDayPL}
\title{Get daily gains/losses.}
\usage{
rh_getDayPL(userAuthToken, accountNumber)
}
\arguments{
\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}

\item{accountNumber}{account number generated by rh_getAccountNumber.}
}
\description{
Returns daily change in portfolio value
}
