% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_getFundamentals}
\alias{rh_getFundamentals}
\title{Get fundamentals}
\usage{
rh_getFundamentals(ticker)
}
\arguments{
\item{ticker}{Equity ticker symbol. All caps string.}
}
\description{
Get summary fundamentals data on company ticker.
}
