% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robinhoodR6.R
\name{rh_getInstrumentInfo}
\alias{rh_getInstrumentInfo}
\title{Get instrument info.}
\usage{
rh_getInstrumentInfo(instrumentURL)
}
\arguments{
\item{instrumentURL}{instrument url returned by rh_getInstrumentURL.}
}
\description{
Return list with symbol and name for instrument url
}
