% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_getInstrumentURL}
\alias{rh_getInstrumentURL}
\title{Get instrument url.}
\usage{
rh_getInstrumentURL(ticker, userAuthToken)
}
\arguments{
\item{ticker}{equity ticker symbol. Must be capitalized.}

\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}
}
\description{
Helper method to get a ticker's instrument url for API requests.
}
