% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_getOptionsOrders}
\alias{rh_getOptionsOrders}
\title{Get option order history}
\usage{
rh_getOptionsOrders(userAuthToken)
}
\arguments{
\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}
}
\description{
Retrieve all options transation history in data.frame
}
