% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_getOptionsPositions}
\alias{rh_getOptionsPositions}
\title{Get options positions}
\usage{
rh_getOptionsPositions(userAuthToken, accountNumber)
}
\arguments{
\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}

\item{accountNumber}{account number generated by rh_getAccountNumber.}
}
\description{
Return list of account's current options positions.
}
