% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_getPortfolioEquity}
\alias{rh_getPortfolioEquity}
\title{Get portfolio equity.}
\usage{
rh_getPortfolioEquity(userAuthToken, accountNumber)
}
\arguments{
\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}

\item{accountNumber}{account number generated by rh_getAccountNumber.}
}
\description{
Returns total portfolio value (equity, cash, and options)
}
