% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_getPositions}
\alias{rh_getPositions}
\title{Get current positions.}
\usage{
rh_getPositions(userAuthToken, accountNumber)
}
\arguments{
\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}

\item{accountNumber}{account number generated by rh_getAccountNumber.}
}
\description{
Return list of account's current positions in list structure of request.
}
