% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_getRecentOrders}
\alias{rh_getRecentOrders}
\title{Get order history}
\usage{
rh_getRecentOrders(userAuthToken)
}
\arguments{
\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}
}
\description{
Retrieve all transation history in data.frame
}
