% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robinhoodR6.R
\name{rh_historicals_one}
\alias{rh_historicals_one}
\title{Historical data helper.}
\usage{
rh_historicals_one(symbol, interval = interval, span = span,
  bounds = bounds, account = account)
}
\arguments{
\item{symbol}{The shorthand symbols.}

\item{interval}{The interval: week|day|10minute|5minute}

\item{span}{The span: day|week|year}

\item{bounds}{extended|regular|trading}

\item{account}{Initialized instance of the R6 Account class. Generated by Account$new(login).
Base url}
}
\description{
Historical data helper.
}
