% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robinhoodR6.R
\name{rh_placeImmediateBuyOrder}
\alias{rh_placeImmediateBuyOrder}
\title{Place immediate buy order.}
\usage{
rh_placeImmediateBuyOrder(account, ticker, orderType, orderPrice,
  numShares)
}
\arguments{
\item{account}{Initialized instance of the R6 Account class. Generated by Account$new(login).}

\item{ticker}{Equity ticker symbol. All caps string.}

\item{orderType}{"market" or "limit"}

\item{orderPrice}{numeric limit price (needed for market order because collared on price)}

\item{numShares}{number of shares. Numeric class integer only.}
}
\description{
Place an immediate trigger market/limit buy order.
}
