% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robinhoodR6.R
\name{rh_placeImmediateSellOrder}
\alias{rh_placeImmediateSellOrder}
\title{Place immediate sell order.}
\usage{
rh_placeImmediateSellOrder(account, ticker, orderType, orderPrice,
  numShares)
}
\arguments{
\item{ticker}{Equity ticker symbol. All caps string.}

\item{orderType}{"market" or "limit"}

\item{orderPrice}{numeric limit price (needed for market order because collared on price)}

\item{numShares}{number of shares. Numeric class integer only.}

\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}

\item{accountNumber}{8-digit account number generated by rh_getAccountNumber.}
}
\description{
Place an immediate trigger market/limit sell order.
}
