% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robinhoodR6.R
\docType{data}
\name{Account}
\alias{Account}
\title{Account R6 Class}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Account
}
\arguments{
\item{login}{Initialized R6 Login class instance.}
}
\description{
Create an instance of the R6 Account class.
}
\keyword{datasets}
