% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robinhoodR6.R
\docType{data}
\name{Login}
\alias{Login}
\title{get r AuthToken.}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Login
}
\arguments{
\item{username_}{Robinhood login username. NOT email.}

\item{password_}{Robinhood login password. Can only contain letters and numbers in plaintext.}
}
\description{
Get oauth2 login token. Passed to all API requests.
}
\keyword{datasets}
