% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robinhoodR6.R
\name{rh_getInstrumentURL}
\alias{rh_getInstrumentURL}
\title{Get instrument url.}
\usage{
rh_getInstrumentURL(ticker, account)
}
\arguments{
\item{ticker}{equity ticker symbol. Must be capitalized.}

\item{account}{Initialized instance of the R6 Account class. Generated by Account$new(login).}
}
\description{
Helper method to get a ticker's instrument url for API requests.
}
