% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_historicals}
\alias{rh_historicals}
\title{Get the historical quotes for a symbol.}
\usage{
rh_historicals(symbols, interval = "day", span = "year",
  bounds = "regular", keep_meta = FALSE, to_xts = TRUE,
  userAuthToken = userAuthToken)
}
\arguments{
\item{symbols}{The shorthand symbols.}

\item{interval}{The interval: week|day|10minute|5minute}

\item{span}{The span: day|week|year}

\item{bounds}{extended|regular|trading}

\item{keep_meta}{whether to keep meta data. Defaults to FALSE.}

\item{to_xts}{whether to convert to xts. Defaults to TRUE.}

\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}
}
\description{
Get the historical quotes for a symbol.
}
