% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rh_funcs.R
\name{rh_placeImmediateBuyOrder}
\alias{rh_placeImmediateBuyOrder}
\title{Place immediate buy order.}
\usage{
rh_placeImmediateBuyOrder(userAuthToken, accountNumber, ticker, orderType,
  orderPrice, numShares)
}
\arguments{
\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}

\item{accountNumber}{8-digit account number generated by rh_getAccountNumber.}

\item{ticker}{Equity ticker symbol. All caps string.}

\item{orderType}{"market" or "limit"}

\item{orderPrice}{numeric limit price (needed for market order because collared on price)}

\item{numShares}{number of shares. Numeric class integer only.}
}
\description{
Place an immediate trigger market/limit buy order.
}
