% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robinhoodR6.R
\name{rh_placeStopBuyOrder}
\alias{rh_placeStopBuyOrder}
\title{Place stop buy order.}
\usage{
rh_placeStopBuyOrder(account, ticker, orderType, orderPrice, numShares,
  stopPrice)
}
\arguments{
\item{ticker}{Equity ticker symbol. All caps string.}

\item{orderType}{"market" or "limit"}

\item{orderPrice}{numeric limit price (needed for market order because collared on price)}

\item{numShares}{number of shares. Numeric class integer only.}

\item{stopPrice}{numeric stop price}

\item{userAuthToken}{oauth2 token generated by rh_getAuthToken.}

\item{accountNumber}{8-digit account number generated by rh_getAccountNumber.}
}
\description{
Place a stop trigger market/limit buy order.
}
