% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robinhoodR6.R
\name{rh_quote}
\alias{rh_quote}
\title{Get the last quotes for a symbol.}
\usage{
rh_quote(ticker, header)
}
\arguments{
\item{header}{Authentication header generated by instance of the R6 Login class. login$authHeader.}

\item{symbols}{The shorthand symbols.}
}
\description{
Get the last quotes for a symbol.
}
